Moved applicatoin classes
