+++ lib/Container/CoreExtension.php
@@ -7,15 +7,15 @@ use PhpBench\DependencyInjection\Container;
 use PhpBench\DependencyInjection\ExtensionInterface;
 use Symfony\Component\Console\Application;
 use Phpactor\UserInterface\Console\Command\ClassMoveCommand;
-use Phpactor\Application\ClassMover\ClassMover as ClassMoverApp;
+use Phpactor\Application\ClassMover as ClassMoverApp;
 use DTL\Filesystem\Adapter\Git\GitFilesystem;
 use DTL\Filesystem\Domain\Cwd;
 use DTL\ClassMover\ClassMover;
 use DTL\Filesystem\Adapter\Simple\SimpleFilesystem;
-use Phpactor\Application\FileInfo\FileInfo;
+use Phpactor\Application\FileInfo;
 use DTL\TypeInference\TypeInference;
 use Phpactor\UserInterface\Console\Command\FileInfoAtOffsetCommand;
-use Phpactor\Application\ClassSearch\ClassSearch;
+use Phpactor\Application\ClassSearch;
 use Phpactor\UserInterface\Console\Command\ClassSearchCommand;
 use DTL\Filesystem\Adapter\Composer\ComposerFilesystem;
 use DTL\Filesystem\Domain\FilePath;
